requirements.txt
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-writable-nested==0.7.2
psycopg2==2.9.10
pytz==2025.2
reportlab==4.4.4